refactor(benchmarks): isolate harness surfaces
