+++ apps/benchmarks/src/surfaces/conformance/conformance-surface.boundary.test.ts
@@ -3,17 +3,18 @@ import { describe, expect, it } from 'vitest';
 
 describe('conformance surface boundaries', () => {
   it('keeps route composition separate from retained and finite renderer ownership', async () => {
-    const [app, surface, comparisonSurface, finiteCapture, finiteSurface] = await Promise.all([
-      readFile(new URL('../../app.tsx', import.meta.url), 'utf8'),
+    const [controller, harnessScene, surface, comparisonSurface, finiteCapture, finiteSurface] = await Promise.all([
+      readFile(new URL('../../controllers/harness-controller.tsx', import.meta.url), 'utf8'),
+      readFile(new URL('../harness/scene.tsx', import.meta.url), 'utf8'),
       readFile(new URL('./conformance-surface.tsx', import.meta.url), 'utf8'),
       readFile(new URL('./raster-technique-comparison-surface.tsx', import.meta.url), 'utf8'),
       readFile(new URL('./capture.ts', import.meta.url), 'utf8'),
       readFile(new URL('./finite-conformance-surface.tsx', import.meta.url), 'utf8'),
     ]);
 
-    expect(app).toContain("from './surfaces/conformance/conformance-surface'");
-    expect(app).not.toContain('function ConformanceSurface(');
-    expect(app).not.toContain('function captureFiniteConformance(');
+    expect(controller).not.toContain('function ConformanceSurface(');
+    expect(controller).not.toContain('function captureFiniteConformance(');
+    expect(harnessScene).toContain("import { ConformanceSurface } from '../conformance/conformance-surface';");
     expect(surface).toContain('<RasterTechniqueComparisonSurface');
     expect(surface).toContain('<FiniteConformanceSurface');
     expect(comparisonSurface).toContain("import('./scenes/raster-technique-comparison')");
